+++ packages/iov-bcp/src/atomicswaphelpers.ts
@@ -1,18 +1,15 @@
 import { Random, Sha256 } from "@iov/crypto";
 
-import { Preimage } from "./atomicswaptypes";
+import { Hash, Preimage } from "./atomicswaptypes";
 
 export class AtomicSwapHelpers {
   public static async createPreimage(): Promise<Preimage> {
-    // We use 16 bytes of random data (128 bit entropy)
-    // which is enough to avoid brute force attacks
-    // given the short lifetime of an atomic swap.
-    const bytes = await Random.getBytes(16);
+    const bytes = await Random.getBytes(32);
     return bytes as Preimage;
   }
 
   /** Creates a SHA256 hash of the preimage */
-  public static hashPreimage(preimage: Preimage): Uint8Array {
-    return new Sha256(preimage).digest();
+  public static hashPreimage(preimage: Preimage): Hash {
+    return new Sha256(preimage).digest() as Hash;
   }
 }