+++ packages/iov-bcp/types/atomicswaptypes.d.ts
@@ -2,11 +2,13 @@ import { As } from "type-tagger";
 import { Stream } from "xstream";
 import { BcpConnection } from "./connection";
 import { Address, Amount, SwapIdBytes, SwapTimeout } from "./transactions";
-export declare enum SwapState {
+export declare enum SwapProcessState {
     Open = "open",
     Claimed = "claimed",
     Aborted = "aborted"
 }
+export declare type Preimage = Uint8Array & As<"preimage">;
+export declare type Hash = Uint8Array & As<"hash">;
 export interface SwapData {
     readonly id: SwapIdBytes;
     readonly sender: Address;
@@ -16,7 +18,7 @@ export interface SwapData {
      *
      * Until we have a way to specify the hashing algirithm, this is SHA256.
      */
-    readonly hash: Uint8Array;
+    readonly hash: Hash;
     readonly amounts: ReadonlyArray<Amount>;
     /**
      * The first point in time at which the offer is expired.
@@ -27,18 +29,17 @@ export interface SwapData {
     readonly memo?: string;
 }
 export interface OpenSwap {
-    readonly kind: SwapState.Open;
+    readonly kind: SwapProcessState.Open;
     readonly data: SwapData;
 }
-export declare type Preimage = Uint8Array & As<"preimage">;
 export interface ClaimedSwap {
-    readonly kind: SwapState.Claimed;
+    readonly kind: SwapProcessState.Claimed;
     readonly data: SwapData;
     readonly preimage: Preimage;
 }
 /** A swap offer that has been aborted */
 export interface AbortedSwap {
-    readonly kind: SwapState.Aborted;
+    readonly kind: SwapProcessState.Aborted;
     readonly data: SwapData;
 }
 export declare type AtomicSwap = OpenSwap | ClaimedSwap | AbortedSwap;